refactor and creation of IntellijPluginInstaller
